pyproject.toml
[project]
name = "repo_scanner"
version = "0.1.0"
description = ""
authors = [
    {name = "krathee95",email = "kartik.rathee.eee14@itbhu.ac.in"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-yasg (>=1.21.9,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
